Add files via upload to fix error
Putting the original files, before preprocessing, into the main folder as these are used in the code to check whether a study is a map or peaks. If run before this correction, demo data code would return an error "Error using adjusting_FE_RE_and_ME (line 243)
Cannot determine whether file Aridan_PrePrint is an image or text file" - apologies for this.